refactor: extract inline route handlers into controllers
Routes were acting as both route and controller — defining business
logic, error handling, and response formatting inline. This violates
the Controllers → Services → Models architecture documented in CLAUDE.md.

- Create 5 new controllers: session, device-token, application-draft,
  foster, invitation
- Slim route files to only validation rules + middleware + controller calls
- Fix admin.routes.ts reminder endpoint to use instanceof ApiError
  instead of string matching
- Fix discovery.routes.ts to stop leaking error.message in 500 response
- Change legal-reminder 'no email' error to UnprocessableError (422)
  to preserve the original semantics
