+++ service.backend/src/routes/invitation.routes.ts
@@ -1,16 +1,10 @@
 import { Router, Request, Response } from 'express';
-import { body, param, validationResult } from 'express-validator';
+import { body, param } from 'express-validator';
 import { sensitiveWriteLimiter } from '../middleware/rate-limiter';
-import { ApiError } from '../middleware/error-handler';
-import { InvitationService } from '../services/invitation.service';
-import User from '../models/User';
-import { logger } from '../utils/logger';
+import { InvitationController } from '../controllers/invitation.controller';
 
 const router = Router();
 
-// Accept invitation (public route - no authentication required)
-// ADS-458: tighter limit on this destructive endpoint to discourage
-// brute-forcing tokens / creating accounts in bulk.
 router.post(
   '/accept',
   sensitiveWriteLimiter,
@@ -31,129 +25,13 @@ router.post(
       .isLength({ max: 100 })
       .withMessage('Title must be max 100 characters'),
   ],
-  async (req: Request, res: Response) => {
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          success: false,
-          message: 'Validation failed',
-          errors: errors.array(),
-        });
-      }
-
-      const { token, firstName, lastName, password, title } = req.body;
-
-      const result = await InvitationService.acceptInvitation(token, {
-        firstName,
-        lastName,
-        password,
-        title,
-      });
-
-      res.status(201).json(result);
-    } catch (error) {
-      logger.error('Error accepting invitation:', {
-        error: error instanceof Error ? error.message : String(error),
-      });
-
-      if (error instanceof ApiError) {
-        return res.status(error.statusCode).json({
-          success: false,
-          message: error.message,
-        });
-      }
-
-      res.status(500).json({
-        success: false,
-        message: 'Failed to accept invitation',
-      });
-    }
-  }
+  (req: Request, res: Response) => InvitationController.acceptInvitation(req, res)
 );
 
-// Get invitation details (public route for form pre-population).
-// The catch wrapper is intentional — `req.originalUrl` contains the
-// invitation token, so we must not let the error propagate to the
-// central error handler (which logs `req.originalUrl`) without
-// scrubbing. The catch logs only `err.message` and emits a generic
-// 500, keeping the raw token out of structured logs.
 router.get(
   '/details/:token',
   [param('token').notEmpty().withMessage('Invitation token is required')],
-  async (req: Request, res: Response) => {
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          success: false,
-          message: 'Validation failed',
-          errors: errors.array(),
-        });
-      }
-
-      const { token } = req.params;
-
-      // Get invitation details without sensitive information
-      const invitation = await InvitationService.getInvitationDetails(token as string);
-
-      if (!invitation) {
-        return res.status(404).json({
-          success: false,
-          message: 'Invitation not found or expired',
-        });
-      }
-
-      // C2-4: surface rescue / inviter / role so the AcceptInvitation
-      // page can show the invitee who invited them and where to. The
-      // `rescue` is an eager-loaded include on the service side; the
-      // `invited_by` column stores a userId with no association on
-      // that FK, so we resolve the name with a separate lookup.
-      const invitationWithRescue = invitation as typeof invitation & {
-        rescue?: { name?: string | null } | null;
-      };
-      const rescueName = invitationWithRescue.rescue?.name ?? null;
-
-      let invitedByName: string | null = null;
-      if (invitation.invited_by) {
-        const inviter = await User.findOne({
-          where: { userId: invitation.invited_by },
-          attributes: ['firstName', 'lastName'],
-        });
-        if (inviter) {
-          const fullName = `${inviter.firstName} ${inviter.lastName}`.trim();
-          invitedByName = fullName.length > 0 ? fullName : null;
-        }
-      }
-
-      res.json({
-        success: true,
-        invitation: {
-          email: invitation.email,
-          expiresAt: invitation.expiration,
-          rescueName,
-          invitedByName,
-          role: invitation.title ?? null,
-        },
-      });
-    } catch (error) {
-      logger.error('Error getting invitation details:', {
-        error: error instanceof Error ? error.message : String(error),
-      });
-
-      if (error instanceof ApiError) {
-        return res.status(error.statusCode).json({
-          success: false,
-          message: error.message,
-        });
-      }
-
-      res.status(500).json({
-        success: false,
-        message: 'Failed to get invitation details',
-      });
-    }
-  }
+  (req: Request, res: Response) => InvitationController.getDetails(req, res)
 );
 
 export default router;
